requirements.txt
psycopg2-binary
django
django-extensions
django-crispy-forms
django-widget-tweaks
runscript
gunicorn
python-dateutil
pytz
pymongo
ujson
pyyaml
